requirements.txt
# Core Dependencies
rich>=10.0.0
typer>=0.4.0
pydantic>=1.8.2
asyncio>=3.4.3
aiohttp>=3.8.1
cryptography>=3.4.7
PyYAML>=6.0
python-dotenv>=0.19.0

# Security Tools
scapy>=2.4.5
pyOpenSSL>=20.0.1
paramiko>=2.8.1
requests>=2.26.0
beautifulsoup4>=4.9.3
dnspython>=2.1.0
netfilterqueue>=1.0.0
python-nmap>=0.7.1
pymetasploit3>=1.0.3
impacket>=0.9.24

# AI/ML Components
tensorflow>=2.7.0
torch>=1.10.0
scikit-learn>=1.0.1
numpy>=1.21.4
pandas>=1.3.4
transformers>=4.12.5

# System Integration
psutil>=5.8.0
docker>=5.0.3
kubernetes>=18.20.0
prometheus-client>=0.12.0
ray[serve]>=1.8.0
optuna>=2.10.0

# Web Security
sqlmap-python>=1.5.12
wfuzz>=3.1.0
wafw00f>=2.1.0
dirbuster>=1.0.0
nikto>=2.1.6

# Network Analysis
wireshark-python>=3.4.9
pyshark>=0.4.5
pcapy>=0.11.5
dpkt>=1.9.7
pypcap>=1.2.3

# Forensics
volatility3>=2.0.1
sleuthkit>=4.10.0
autopsy>=4.19.0
foremost>=1.5.7
scalpel>=2.0.0

# Cryptography
pycryptodome>=3.11.0
hashlib>=20.3.0
bcrypt>=3.2.0
passlib>=1.7.4
jwt>=1.3.1

# Development Tools
pytest>=6.2.5
black>=21.12b0
mypy>=0.910
pylint>=2.12.2
coverage>=6.2

# Documentation
sphinx>=4.3.2
mkdocs>=1.2.3
pdoc3>=0.10.0
jupyter>=1.0.0
nbconvert>=6.3.0

# Monitoring & Logging
loguru>=0.5.3
sentry-sdk>=1.5.1
elastic-apm>=6.7.2
statsd>=3.3.0
datadog>=0.43.0

# UI Components
textual>=0.1.18
prompt_toolkit>=3.0.24
click>=8.0.3
tqdm>=4.62.3
colorama>=0.4.4

# Testing & Quality
bandit>=1.7.1
safety>=1.10.3
pytype>=2021.12.15
semgrep>=0.75.0
pre-commit>=2.16.0 